refact: move lib/my/* (except my/img) to lib-old/my/
