+++ app/lib/my-img/node-magick.js
